--- src/types/meme-comment.ts
@@ -1,7 +1,0 @@
-export type MemeComment = {
-  id: string;
-  authorId: string;
-  memeId: string;
-  content: string;
-  createdAt: string;
-};
