--- src/types/meme-text.ts
@@ -1,5 +1,0 @@
-export type MemeText = {
-  content: string;
-  x: number;
-  y: number;
-};
