--- src/types/pagination-response.ts
@@ -1,5 +1,0 @@
-export type PaginationResponse<T> = {
-  total: number;
-  pageSize: number;
-  results: T[];
-};
