--- src/types/user.ts
@@ -1,5 +1,0 @@
-export type User = {
-  id: string;
-  username: string;
-  pictureUrl: string;
-};
